Update yay installation for Arch Linux
`yay` can not be installed using `pacman -Syu yay` because `yay` is not in the Arch repository but only in the Arch User Repository (AUR).

The revised instructions provide two methods: one for compiling `yay` from source and another for downloading the pre-built binary version. Also included a link to the yay repository for more detailed instructions.